+++ docs/history.rst
@@ -129,9 +129,10 @@ New Features
   the *mem* and *obj* domains as well as ``pymalloc``'s arena allocator.
   (However, the ``pymalloc`` arena allocator customization does not yet work
   due to a bug outside of our control.)
-* The *mimalloc* memory allocator can now be used as Python's *raw allocator*.
-  See documentation for
-  :ref:`PythonInterpreterConfig.raw_allocator <config_type_python_interpreter_config_allocator_backend>`.
+* The *mimalloc* and *snmalloc* memory allocators can now be used as Python's
+  memory allocators. See documentation for
+  :ref:`PythonInterpreterConfig.allocator_backend <config_type_python_interpreter_config_allocator_backend>`.
+  Code contributed by Ryan Clanton in #358.
 
 Other Relevant Changes
 ^^^^^^^^^^^^^^^^^^^^^^